+++ deployment/deploy.sh
@@ -23,6 +23,8 @@ LIGHTCYAN='\033[1;36m'
 WHITE='\033[1;37m'
 DEFAULT_COLOR=${LIGHTBLUE}
 
+LOCAL=false
+
 get_internal_node_ip() {
   kubectl get nodes -l "$1" -o jsonpath='{.items[*].status.addresses[?(@.type=="InternalIP")].address}'
 }
@@ -46,13 +48,33 @@ msg() {
   echo -e "${DEFAULT_COLOR}$1${NOCOLOR}"
 }
 
+###############################################################################
+# Start of the actual script                                                  #
+###############################################################################
+
+usage() {
+  echo "Usage: $0 [-l]" 1>&2; exit 1;
+}
+
+while getopts "l" flag; do
+  case "${flag}" in
+    l)
+      LOCAL=true ;;
+    ?)
+      usage ;;
+  esac
+done
 
 ###############################################################################
 # Start of the actual script                                                  #
 ###############################################################################
 
 big-sep
-msg "| Starting local deployment 🧙‍"
+if [[ $LOCAL == true ]]; then
+  msg "| Starting local deployment 🧙‍"
+else
+  msg "| Starting deployment 🧙‍"
+fi
 big-sep
 
 # Create fresh kind cluster
@@ -60,31 +82,29 @@ header "Create fresh kind cluster"
 kind delete clusters kind-t2
 kind create cluster --name kind-t2 --config deployment/kind-t2-config.yaml
 
-# Build and load docker images of services
-header "Build and load images of services"
-set -e
-./gradlew clean build
-./gradlew ":credit:bootBuildImage"
-./gradlew ":issue:bootBuildImage"
-./gradlew ":basketserver:bootBuildImage"
-kind load docker-image upbcuk/incentive-service-issue --name kind-t2
-kind load docker-image upbcuk/incentive-service-credit --name kind-t2
-kind load docker-image upbcuk/incentive-service-basketserver --name kind-t2
-set +e
+if [[ $LOCAL == true ]]; then
+  # Build and load docker images of services
+  header "Build and load images of services"
+  set -e
+  ./gradlew clean build
+  ./gradlew ":credit:bootBuildImage"
+  ./gradlew ":issue:bootBuildImage"
+  ./gradlew ":basketserver:bootBuildImage"
+  kind load docker-image upbcuk/incentive-service-issue --name kind-t2
+  kind load docker-image upbcuk/incentive-service-credit --name kind-t2
+  kind load docker-image upbcuk/incentive-service-basketserver --name kind-t2
+  set +e
 
-# Print available images
-header "Available container images:"
-docker exec -it kind-t2-control-plane crictl images
+  # Print available images
+  header "Available container images:"
+  docker exec -it kind-t2-control-plane crictl images
+fi
 
 # Deploy to kubernetes
 # Create ingress
 header "Create ingress"
-kubectl apply -f https://raw.githubusercontent.com/kubernetes/ingress-nginx/master/deploy/static/provider/kind/deploy.yaml
-kubectl wait --namespace ingress-nginx \
-  --for=condition=ready pod \
-  --selector=app.kubernetes.io/component=controller \
-  --timeout=300s
 kubectl apply -f deployment/ingress.yaml
+kubectl apply -f https://raw.githubusercontent.com/kubernetes/ingress-nginx/master/deploy/static/provider/kind/deploy.yaml
 
 # Deploy services
 header "Deploy services"
@@ -96,4 +116,5 @@ echo ""
 big-sep
 msg "| Done! 🥳"
 msg "| Node-ip: $(get_internal_node_ip "cuk.cs.upb.de/kind-control-plane")"
+msg "| It may take some time until all services are reachable"
 big-sep