+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0] - WIP
+## Next Release
+ - Add options for local deployment and deployment using the container images from dockerhub and fix ingress configuration in `deploy.sh` [#8](https://github.com/upbcuk/incentive-services/pull/8)
+ - Fix wrong issue service configuration bug [#9](https://github.com/upbcuk/incentive-services/pull/9)
  - Add basket server [#10](https://github.com/upbcuk/incentive-services/pull/10)
 
 ## [0.3.0] - 2021-01-12